largest prime

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,8 +9,8 @@
 
 int main(void)
 {
-	int i;
-	int prime;
+	long long i;
+	long long  prime;
 	long long n = 612852475143;
 
 	for (i = 2; i <= n; i++)
